+++ __tests__/__src__/cli/submit/local-file/LocalFile.Handler.test.ts
@@ -57,6 +57,7 @@ describe("Submit job from local file handler", () => {
             owner: "owner2",
             status: "OUTPUT",
             rc: 0,
+            retcode: "RC 0000",
             spoolFiles: [
                 {
                     name: "file1"